load_balancer/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==6.1.0